fix: left_count to float

+++ src/models/cards/mod.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct CardBody {
     pub title: String,
     pub exp_date: String,
-    pub left_count: i32,
+    pub left_count: f64,
     pub units: String,
     pub list_id: i32,
     pub user_id: i32,
@@ -15,7 +15,7 @@ pub struct Card {
     pub id: i32,
     pub title: String,
     pub exp_date: chrono::NaiveDateTime,
-    pub left_count: i32,
+    pub left_count: f64,
     pub units: String,
     pub list_id: i32,
     pub user_id: i32,
